* Use 64 bits length in tlb miss exception where required. * Support 9x range in 64 bits space.

+++ Exception.h
@@ -70,7 +70,7 @@ void DoIntegerOverflow					( BOOL DelaySlot );
 void DoAddressError						( BOOL DelaySlot, QWORD BadVaddr, BOOL FromRead );
 void _fastcall DoCopUnusableException	( BOOL DelaySlot, int Coprocessor );
 void DoIntrException					( BOOL DelaySlot );
-void _fastcall DoTLBMiss				( BOOL DelaySlot, DWORD BadVaddr, BOOL FromRead );
+void _fastcall DoTLBMiss				( BOOL DelaySlot, QWORD BadVaddr, BOOL FromRead );
 void _fastcall DoSysCallException		( BOOL DelaySlot );
 void _fastcall DoBreakException			( BOOL DelaySlot );
 void _fastcall DoIllegalInstructionException(BOOL DelaySlot );
